Fix up negative binomial table

+++ src/main/java/com/bromleyoil/binomialstats/model/Range.java
@@ -1,10 +1,10 @@
 package com.bromleyoil.binomialstats.model;
 
 public class Range {
-	private long x1;
-	private long x2;
+	private int x1;
+	private int x2;
 
-	public Range(long x1, long x2) {
+	public Range(int x1, int x2) {
 		this.x1 = x1;
 		this.x2 = x2;
 	}
